% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{as.ecdf.default}
\alias{as.ecdf.default}
\title{Default implementation of the as.ecdf generic, to provide backwards compatibility}
\usage{
\method{as.ecdf}{default}(x)
}
\description{
Default implementation of the as.ecdf generic, to provide backwards compatibility
}
