% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{as.ecdf.hist_dat}
\alias{as.ecdf.hist_dat}
\title{Convert this histogram to an instance of the "ecdf" class, allowing the
calculation of cumulative densities, and quantiles}
\usage{
\method{as.ecdf}{hist_dat}(x)
}
\arguments{
\item{x}{An instance of the class his_dat}
}
\description{
Convert this histogram to an instance of the "ecdf" class, allowing the
calculation of cumulative densities, and quantiles
}
