% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{as.vector.default}
\alias{as.vector.default}
\title{Default implementation of the as.vector generic, to provide backwards compatibility}
\usage{
\method{as.vector}{default}(x)
}
\arguments{
\item{x}{The object to convert to a vector}
}
\description{
Default implementation of the as.vector generic, to provide backwards compatibility
}
