% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{as.vector.hist_dat}
\alias{as.vector.hist_dat}
\title{Convert this histogram to a vector. Not recommended if there are many counts
as this would result in an incredibly long vector}
\usage{
\method{as.vector}{hist_dat}(his, ...)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\description{
Convert this histogram to a vector. Not recommended if there are many counts
as this would result in an incredibly long vector
}
