% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{as.vector}
\alias{as.vector}
\title{Converts an object to a vector.
This is a generic version of the core \code{as.vector} function}
\usage{
as.vector(x)
}
\arguments{
\item{x}{The object to coerce to a vector}
}
\value{
A vector
}
\description{
Converts an object to a vector.
This is a generic version of the core \code{as.vector} function
}
