% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{c,HistDatCompatible-method}
\alias{c,HistDatCompatible-method}
\title{Concatenate observations into this instance}
\usage{
\S4method{c}{HistDatCompatible}(x, ...)
}
\arguments{
\item{HistDat}{}
}
\value{

}
\description{
Concatenate observations into this instance
}
