% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{hist_dat}
\alias{hist_dat}
\title{S3 class for histogram data}
\usage{
hist_dat(vals, counts)
}
\arguments{
\item{vals}{A vector of observations}

\item{counts}{A vector of counts, each of which corresponds to the same
index in the vals parameter}
}
\value{
An instance of the hist_dat class
}
\description{
S3 class for histogram data
}
