% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{HistDat}
\alias{HistDat}
\title{The constructor function for the HistDat class. This is the only official way
to create an instance of this class.}
\usage{
HistDat(vals, counts)
}
\arguments{
\item{vals}{A vector of observation values, ie all the possible values that
could be observed}

\item{counts}{A vector of counts, each of which corresponds to the same
index in the vals parameter}
}
\description{
The constructor function for the HistDat class. This is the only official way
to create an instance of this class.
}
\examples{
hd <- HistDat::HistDat(vals = 1:3, counts = c(1, 2, 1)) # equivalent to above
length(hd) # returns 4
}
