% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{length,HistDat-method}
\alias{length,HistDat-method}
\title{The number of values in a histogram dataset}
\usage{
\S4method{length}{HistDat}(x)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments that will be ignored}
}
\value{
A numeric of length 1
}
\description{
The number of values in a histogram dataset
}
\examples{
length(new("HistDat", vals=1:3, counts=c(1, 2, 3)))
}
