% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{length.hist_dat}
\alias{length.hist_dat}
\title{The number of values in a histogram dataset}
\usage{
\method{length}{hist_dat}(his, na.rm)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The number of values in a histogram dataset
}
