% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{max,HistDat-method}
\alias{max,HistDat-method}
\title{The maximum value in a histogram dataset}
\usage{
\S4method{max}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{max()}}
}
\value{
A numeric of length 1
}
\description{
The maximum value in a histogram dataset
}
\examples{
max(new("HistDat", vals=1:3, counts=c(1, 2, 3)))
}
