% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{max.hist_dat}
\alias{max.hist_dat}
\title{The maximum value in a histogram dataset}
\usage{
\method{max}{hist_dat}(his, na.rm)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The maximum value in a histogram dataset
}
