% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{mean,HistDat-method}
\alias{mean,HistDat-method}
\title{The mean value of all observations in the histogram dataset}
\usage{
\S4method{mean}{HistDat}(x, ...)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{...}{Additional arguments that will be ignored}
}
\value{
A numeric of length 1, holding the mean of the observationsin the
dataset
}
\description{
The mean value of all observations in the histogram dataset
}
\examples{
hd = HistDat(vals=1:3, counts=c(1, 2, 1))
mean(hd) # returns 2
}
