% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{mean.hist_dat}
\alias{mean.hist_dat}
\title{The mean value in a histogram dataset}
\usage{
\method{mean}{hist_dat}(his)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The mean value in a histogram dataset
}
