% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{median,HistDat-method}
\alias{median,HistDat-method}
\title{Calculates the median value of the observations in the histogram dataset}
\usage{
\S4method{median}{HistDat}(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{na.rm}{Provided for compatibility with \code{\link[stats:median]{stats::median()}}, but ignored}

\item{...}{Additional arguments that will be ignored}
}
\value{
A numeric of length 1, holding the median value of the observations
in the histogram dataset
}
\description{
Calculates the median value of the observations in the histogram dataset
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
median(hd) # returns 2
}
