% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{min,HistDat-method}
\alias{min,HistDat-method}
\title{The minimum value in a histogram dataset}
\usage{
\S4method{min}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class his_dat
min(new("HistDat", vals=1:3, counts=c(1, 2, 3)))}

\item{...}{Additional arguments to pass to \code{min()}}
}
\value{
A numeric of length 1
}
\description{
The minimum value in a histogram dataset
}
