% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{min.hist_dat}
\alias{min.hist_dat}
\title{The minimum value in a histogram dataset}
\usage{
\method{min}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{min()}}
}
\value{
A numeric of length 1
}
\description{
The minimum value in a histogram dataset
}
