% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{range,HistDat-method}
\alias{range,HistDat-method}
\title{The range of values in a histogram dataset}
\usage{
\S4method{range}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{range()}}
}
\value{
A numeric of length 2, indicating the minimum and maximum value
}
\description{
The range of values in a histogram dataset
}
\examples{
range(new("HistDat", vals=1:3, counts=c(1, 2, 3)))
}
