% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{range.hist_dat}
\alias{range.hist_dat}
\title{The range of values in a histogram dataset}
\usage{
\method{range}{hist_dat}(his, ...)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\value{
A numeric of length 2, indicating the minimum and maximum value
}
\description{
The range of values in a histogram dataset
}
