% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{sd.default}
\alias{sd.default}
\title{Default implementation of the sd generic, to provide backwards compatibility}
\usage{
\method{sd}{default}(x, ...)
}
\description{
Default implementation of the sd generic, to provide backwards compatibility
}
