% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{sd}
\alias{sd}
\title{Calculates the sample standard deviation of a dataset.
This is a generic version of  the core \code{sd} function.}
\usage{
sd(x)
}
\arguments{
\item{x}{The dataset to calculate the standard deviation of.
Can be a vector or any other type implementing the var generic}
}
\value{
A length 1 vector holding the standard deviation
}
\description{
Calculates the sample standard deviation of a dataset.
This is a generic version of  the core \code{sd} function.
}
\examples{
sd(c(1, 2, 3))
}
