% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{[,HistDat,ANY,ANY,ANY-method}
\alias{[,HistDat,ANY,ANY,ANY-method}
\title{Title}
\usage{
\S4method{[}{HistDat,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{HistDat.}
}
\value{

}
\description{
Title
}
