% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{sum,HistDat-method}
\alias{sum,HistDat-method}
\title{Sum of all values in a histogram dataset}
\usage{
\S4method{sum}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{sum()}}
}
\value{
A numeric of length 1
}
\description{
Sum of all values in a histogram dataset
}
\examples{
sum(new("HistDat", vals=1:3, counts=c(1, 2, 3)))
}
