% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{sum.hist_dat}
\alias{sum.hist_dat}
\title{Sum of all values in a histogram dataset}
\usage{
\method{sum}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{sum()}}
}
\value{
A numeric of length 1
}
\description{
Sum of all values in a histogram dataset
}
