% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{var,HistDat-method}
\alias{var,HistDat-method}
\title{The variance of a histogram dataset}
\usage{
\S4method{var}{HistDat}(x)
}
\arguments{
\item{x}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The variance of a histogram dataset
}
\examples{
var(new("HistDat", vals=1:3, counts=c(1, 2, 3)))
}
