% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{var.default}
\alias{var.default}
\title{Default implementation of the var generic, to provide backwards compatibility}
\usage{
\method{var}{default}(x, ...)
}
\description{
Default implementation of the var generic, to provide backwards compatibility
}
