% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{var.hist_dat}
\alias{var.hist_dat}
\title{The variance of a histogram dataset}
\usage{
\method{var}{hist_dat}(x)
}
\arguments{
\item{x}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The variance of a histogram dataset
}
