% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{var}
\alias{var}
\title{Calculates the sample variance of a dataset. This is a generic version of
the core \code{var} function.}
\usage{
var(x, y = NULL, na.rm = FALSE, use)
}
\arguments{
\item{x}{The dataset to calculate the variance of. Can be a vector or any
other type implementing the var generic}
}
\value{
A length 1 vector holding the variance
}
\description{
Calculates the sample variance of a dataset. This is a generic version of
the core \code{var} function.
}
\examples{
var(c(1, 2, 3))
}
