% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\docType{class}
\name{HistDat-class}
\alias{HistDat-class}
\title{S4 class for histogram data}
\description{
S4 class for histogram data
}
\section{Slots}{

\describe{
\item{\code{vals}}{A vector of observations}

\item{\code{counts}}{A vector of counts, each of which corresponds to the same
index in the vals parameter}
}}

