% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\docType{package}
\name{HistDat-package}
\alias{HistDat-package}
\title{'HistDat': Summary statistics for histogram/count data}
\description{
In some cases you will have data in a "histogram" format, where
you have a vector of all possible observations, and a vector of how many
times each observation appeared. You could expand this into a single 1D
vector, but this may not be advisable if the counts are extremely large.
'HistDat' allows for the calculation of summary statistics without the need
for expanding your data.
}
\details{
Note that all the methods described for \code{HistDat} instances have been
transformed into generic methods in this package where they are not already,
with default implementations for general numeric vectors. This allows you
to equally apply these same functions to any type of data.
}
\section{Class Definition / Constructor Function}{
\itemize{
\item \linkS4class{HistDat}
}
}

\section{\code{HistDat} Statistics}{
\itemize{
\item \link{length,HistDat-method}
\item \link{max,HistDat-method}
\item \link{min,HistDat-method}
\item \link{range,HistDat-method},
\item \link{median,HistDat-method}
\item \link{mean,HistDat-method}
\item \link{var,HistDat-method}
\item \link{sd,HistDat-method}
\item \link{sum,HistDat-method}
\item \link{quantile,HistDat-method}
\item \link{as.ecdf,HistDat-method}
}
}

\section{\code{HistDat} Utilities}{
\itemize{
\item \link{as.vector,HistDat-method}
\item \link{[,HistDat-method}
\item \link{c,HistDat-method}
}
}

\section{Misc Functions}{
\itemize{
\item \code{\link[=as.ecdf]{as.ecdf()}}
}
}

