% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{length,HistDat-method}
\alias{length,HistDat-method}
\title{Calculates the total number of observations in a histogram dataset}
\usage{
\S4method{length}{HistDat}(x)
}
\arguments{
\item{x}{An instance of the class HistDat}
}
\value{
A numeric of length 1, holding the number of observations in the
dataset
}
\description{
Calculates the total number of observations in a histogram dataset
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
length(hd) # returns 4
}
