% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{length.hist_dat}
\alias{length.hist_dat}
\title{The number of values in a histogram dataset}
\usage{
\method{length}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments that will be ignored}
}
\value{
A numeric of length 1
}
\description{
The number of values in a histogram dataset
}
