% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{max,HistDat-method}
\alias{max,HistDat-method}
\title{Calculates the largest observation in the histogram dataset}
\usage{
\S4method{max}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{...}{Passed verbatim to \code{\link[base:Extremes]{base::max()}}}

\item{na.rm}{Passed verbatim to \code{\link[base:Extremes]{base::max()}}}
}
\value{
A numeric of length 1, holding the largest observation in the dataset
}
\description{
Calculates the largest observation in the histogram dataset
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
max(hd) # returns 3
}
