% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{max.hist_dat}
\alias{max.hist_dat}
\title{The maximum value in a histogram dataset}
\usage{
\method{max}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{max()}}
}
\value{
A numeric of length 1
}
\description{
The maximum value in a histogram dataset
}
