% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{median.hist_dat}
\alias{median.hist_dat}
\title{The median value in a histogram dataset}
\usage{
\method{median}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments that will be ignored}
}
\value{
A numeric of length 1
}
\description{
The median value in a histogram dataset
}
