% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{min,HistDat-method}
\alias{min,HistDat-method}
\title{Calculates the smallest observation in the histogram dataset}
\usage{
\S4method{min}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{...}{Passed verbatim to \code{\link[base:Extremes]{base::min()}}}

\item{na.rm}{Passed verbatim to \code{\link[base:Extremes]{base::min()}}}
}
\value{
A numeric of length 1, holding the smallest observation in the
dataset
}
\description{
Calculates the smallest observation in the histogram dataset
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
min(hd) # returns 1
}
