% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{min.hist_dat}
\alias{min.hist_dat}
\title{The minimum value in a histogram dataset}
\usage{
\method{min}{hist_dat}(his, na.rm)
}
\arguments{
\item{his}{An instance of the class his_dat}
}
\value{
A numeric of length 1
}
\description{
The minimum value in a histogram dataset
}
