% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{quantile,HistDat-method}
\alias{quantile,HistDat-method}
\title{Returns the empirical quantiles of the observations represented by this
class}
\usage{
\S4method{quantile}{HistDat}(x, ...)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{...}{Remaining arguments to pass to \code{\link[stats:quantile]{stats::quantile()}}}
}
\description{
Returns the empirical quantiles of the observations represented by this
class
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
quantile(hd, 0.1) # returns 1.3
}
