% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{quantile.hist_dat}
\alias{quantile.hist_dat}
\title{Calculates one or more empirical quantiles of the dataset}
\usage{
\method{quantile}{hist_dat}(x, ...)
}
\arguments{
\item{x}{An instance of the class his_dat}

\item{...}{Additional arguments to pass to \code{quantile()}}
}
\value{
A numeric with the same length as the probs parameter
}
\description{
Calculates one or more empirical quantiles of the dataset
}
