% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{range,HistDat-method}
\alias{range,HistDat-method}
\title{Calculates the range of values of the observations in the histogram dataset}
\usage{
\S4method{range}{HistDat}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{...}{Additional arguments to pass to \code{range()}}

\item{na.rm}{Passed verbatim to \code{\link[base:range]{base::range()}}}
}
\value{
A numeric of length 2, indicating the minimum and maximum value of
the observations
}
\description{
Calculates the range of values of the observations in the histogram dataset
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
range(hd) # returns 1 3
}
