% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{sd.default}
\alias{sd.default}
\title{Default implementation of the sd generic, to provide backwards compatibility}
\usage{
\method{sd}{default}(x, ...)
}
\arguments{
\item{x}{The data to calculate the standard deviation of}

\item{...}{Additional arguments to pass to the \code{var} function}
}
\description{
Default implementation of the sd generic, to provide backwards compatibility
}
