% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{[,HistDat,ANY,ANY,ANY-method}
\alias{[,HistDat,ANY,ANY,ANY-method}
\alias{[,HistDat-method}
\alias{sub,HistDat-method}
\title{Index the histogram data}
\usage{
\S4method{[}{HistDat,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An instance of the class HistDat}

\item{i}{A vector of indices to find in the sorted array of observations}

\item{j, drop, ...}{Included for compatibility, but ignored}
}
\value{
The observations that would be returned if you flattened the
array and then indexed it
}
\description{
Index the histogram data
}
\examples{
hd <- HistDat(vals = 1:3, counts = c(1, 2, 1))
hd[1] # returns 1
hd[2] # returns 2
hd[3] # returns 2
}
