% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{var.default}
\alias{var.default}
\title{Default implementation of the var generic, to provide backwards compatibility}
\usage{
\method{var}{default}(x, ...)
}
\arguments{
\item{x}{The data to calculate the variance of}

\item{...}{Additional arguments to pass to \code{stats::var()}}
}
\description{
Default implementation of the var generic, to provide backwards compatibility
}
